byte offset trick
Use byte offset trick to fetch upper bits of the source.
It allows to use exactly the same code for 1 <= n <= 34.
The main advantage of this variant is a consistency of performance
across compilers and ranges.
For the known narrow ranges and the best compiler (i.e. llvm)
it could be somewhat slower than the previous variant or variants,
but it's never much slower.