Another change of algorithm
New algorithm effectively shifts srcH for n=1,2 eith bits to the left.
It allows to select 16 times bigger factors for multiplication by reciproal.
It leads to much more precise evaluation of r1:r2 greatly reducing
probability of execution of 'if (rem >= mulF)'. It means lower probability
of branch misprediction.

A new variant is measurably faster with clang and VisualC compilers.
With gcc - not so much.