New algorithm based on higher-precision multiplication by reciprocal
Calculate product with ~144 bits of precision (~30 good bits of fraction).
Due to such high precision it becomes sufficient to calculate only lower
64 bits of reminder.
Also it allows to use fraction as condition/predictor for underflow,
greatly improving predictability of the following branch.
This versio is significantly faster than the previous one when compiled
with LLVM and a about the same speed as  the previous one when compiled
with VisualC++ (Vs2015).

Unfortunately gcc9 "optimizer" actively subverts all my optimization
efforts by changing the order of evaluation of branch conditions,
so gcc speed is total crap, even relatively to less than great speed
of the previous version under gccn. So it goes.